+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.2.0]
+- Clean up lost menu entries
+- Select container objects from area
+
 ## [2.1.1]
 - Only show areas menu if at least 2 visible items